feat: - project structure modified for testing purposes - added folders in androidTesting folder - created new CardGenerator class - CardTest added
