--- app/src/androidTest/java/at/moritzmusel/gwent/ExampleInstrumentedTest.java
@@ -1,21 +1,0 @@
-package at.moritzmusel.gwent;
-
-import static org.junit.Assert.assertEquals;
-
-import android.content.Context;
-
-import androidx.test.platform.app.InstrumentationRegistry;
-
-import org.junit.Rule;
-import org.junit.Test;
-
-import kotlin.jvm.JvmField;
-
-public class ExampleInstrumentedTest {
-    @Test
-    public void useAppContext() {
-        // Context of the app under test.
-        Context appContext = InstrumentationRegistry.getInstrumentation().getTargetContext();
-        assertEquals("at.moritzmusel.gwent", appContext.getPackageName());
-    }
-}
