+++ app/src/androidTest/java/at/moritzmusel/gwent/ui/LoadingActivityTest.java
@@ -1,4 +1,4 @@
-package at.moritzmusel.gwent;
+package at.moritzmusel.gwent.ui;
 
 import static androidx.test.espresso.Espresso.onView;
 import static androidx.test.espresso.action.ViewActions.click;
@@ -6,6 +6,7 @@ import static androidx.test.espresso.assertion.ViewAssertions.matches;
 import static androidx.test.espresso.matcher.ViewMatchers.isDisplayed;
 import static androidx.test.espresso.matcher.ViewMatchers.withId;
 
+import androidx.test.espresso.matcher.ViewMatchers;
 import androidx.test.ext.junit.rules.ActivityScenarioRule;
 
 import org.junit.After;
@@ -13,14 +14,16 @@ import org.junit.Before;
 import org.junit.Rule;
 import org.junit.Test;
 
-public class SettingsActivityTest {
+import at.moritzmusel.gwent.MainActivity;
+import at.moritzmusel.gwent.R;
+
+public class LoadingActivityTest {
     @Rule
     public ActivityScenarioRule<MainActivity> activityRule = new ActivityScenarioRule<>(MainActivity.class);
 
     @Before
     public void setUp() throws Exception {
-        onView(withId(R.id.imageViewWelcomeScreen)).perform((click()));
-        onView(withId(R.id.imageView_settings)).perform((click()));
+        onView(ViewMatchers.withId(R.id.imageViewWelcomeScreen)).perform((click()));
     }
 
     @After
@@ -28,13 +31,8 @@ public class SettingsActivityTest {
     }
 
     @Test
-    public void test_navigate_SettingsActivity() {
-        //loading
-        try {
-            Thread.sleep(5000);
-        } catch (InterruptedException e) {
-            e.printStackTrace();
-        }
-        //onView(withId(R.id.idFrameLayout)).check(matches(isDisplayed()));
+    public void test_navigate_LoadingActivity() {
+        onView(withId(R.id.textView)).check(matches(isDisplayed()));
+        onView(withId(R.id.progressBar)).check(matches(isDisplayed()));
     }
 }
