+++ app/src/androidTest/java/at/moritzmusel/gwent/ui/WelcomeActivityTest.java
@@ -1,16 +1,20 @@
-package at.moritzmusel.gwent;
+package at.moritzmusel.gwent.ui;
 
 import static androidx.test.espresso.Espresso.onView;
 import static androidx.test.espresso.assertion.ViewAssertions.matches;
 import static androidx.test.espresso.matcher.ViewMatchers.isDisplayed;
 import static androidx.test.espresso.matcher.ViewMatchers.withId;
 
+import androidx.test.espresso.matcher.ViewMatchers;
 import androidx.test.ext.junit.rules.ActivityScenarioRule;
 
 import org.junit.After;
 import org.junit.Before;
 import org.junit.Rule;
 import org.junit.Test;
+
+import at.moritzmusel.gwent.MainActivity;
+import at.moritzmusel.gwent.R;
 
 public class WelcomeActivityTest {
     @Rule
@@ -25,6 +29,6 @@ public class WelcomeActivityTest {
 
     @Test
     public void test_navigate_WelcomeActivity() {
-        onView(withId(R.id.imageViewWelcomeScreen)).check(matches(isDisplayed()));
+        onView(ViewMatchers.withId(R.id.imageViewWelcomeScreen)).check(matches(isDisplayed()));
     }
 }
